+++ tests/test_gemgis.py
@@ -313,10 +313,10 @@ def test_to_gempy_df_lines_data(gdf, dem):
                              rasterio.open('../../gemgis/data/Test1/raster1.tif')
                          ])
 def test_to_gempy_df_lines_xyz_data(gdf, dem):
-    from gemgis.vector import extract_coordinates
+    from gemgis.vector import extract_xyz
     from gemgis import GemPyData
     data = GemPyData(model_name='Model1')
-    gdf_xyz = extract_coordinates(gdf, dem, inplace=True)
+    gdf_xyz = extract_xyz(gdf, dem)
     data.to_gempy_df(gdf_xyz, cat='interfaces')
 
     assert dem.read(1).ndim == 2
@@ -349,10 +349,10 @@ def test_to_gempy_df_lines_xyz_data(gdf, dem):
                              rasterio.open('../../gemgis/data/Test1/raster1.tif')
                          ])
 def test_to_gempy_df_points_xyz_data(gdf, dem):
-    from gemgis.vector import extract_coordinates
+    from gemgis.vector import extract_xyz
     from gemgis import GemPyData
     data = GemPyData(model_name='Model1')
-    gdf_xyz = extract_coordinates(gdf, dem, inplace=True)
+    gdf_xyz = extract_xyz(gdf, dem)
     data.to_gempy_df(gdf_xyz, cat='interfaces')
 
     assert dem.read(1).ndim == 2
@@ -489,8 +489,8 @@ def test_create_surface_color_dict():
 
     data.to_surface_color_dict('../../gemgis/data/Test1/style1.qml')
 
-    assert isinstance(data.surface_color_dict, dict)
-    assert data.surface_color_dict == {'Sand1': '#b35a2a', 'Sand2': '#b35a2a', 'Ton': '#525252'}
+    assert isinstance(data.surface_colors, dict)
+    assert data.surface_colors == {'Sand1': '#b35a2a', 'Sand2': '#b35a2a', 'Ton': '#525252'}
 
 
 def test_create_surface_color_dict_error():
@@ -499,409 +499,6 @@ def test_create_surface_color_dict_error():
 
     with pytest.raises(TypeError):
         data.to_surface_color_dict(['../../gemgis/data/Test1/style1.qml'])
-
-
-# Testing extract_z
-###########################################################
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
-                         ])
-def test_extract_z_points(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=False)
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, rasterio.io.DatasetReader)
-    assert all(gdf_new.geom_type == 'Point')
-
-    assert dem.read(1).ndim == 2
-    assert dem.read(1).shape == (275, 250)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-    assert dem.crs == {'init': 'epsg:4326'}
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'Point')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
-                                            157.81229899479604, 191.31802803451436]
-    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
-                                            719.0939805375339]
-    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
-                                            719.0939805375339]
-
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
-                         ])
-def test_extract_z_points_inplace(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=True)
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, rasterio.io.DatasetReader)
-    assert all(gdf_new.geom_type == 'Point')
-
-    assert dem.read(1).ndim == 2
-    assert dem.read(1).shape == (275, 250)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-    assert dem.crs == {'init': 'epsg:4326'}
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'Point')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
-                                            157.81229899479604, 191.31802803451436]
-    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
-                                            719.0939805375339]
-    assert gdf_new['Z'].head().tolist() == [364.994873046875, 400.3435974121094, 459.54931640625, 525.6910400390625,
-                                            597.6325073242188]
-
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
-                         ])
-def test_extract_z_lines_inplace(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=False)
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, rasterio.io.DatasetReader)
-    assert all(gdf_new.geom_type == 'LineString')
-
-    assert dem.read(1).ndim == 2
-    assert dem.read(1).shape == (275, 250)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-    assert dem.crs == {'init': 'epsg:4326'}
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'LineString')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
-                                            19.150128045807676, 27.79511673965105]
-    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
-                                            310.571692592952]
-    assert gdf_new['Z'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
-                                            310.571692592952]
-
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
-                         ])
-def test_extract_z_lines_inplace(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=True)
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, rasterio.io.DatasetReader)
-    assert all(gdf_new.geom_type == 'LineString')
-
-    assert dem.read(1).ndim == 2
-    assert dem.read(1).shape == (275, 250)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-    assert dem.crs == {'init': 'epsg:4326'}
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'LineString')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
-                                            19.150128045807676, 27.79511673965105]
-    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
-                                            310.571692592952]
-    assert gdf_new['Z'].head().tolist() == [353.9727783203125, 359.03631591796875, 364.28497314453125, 364.994873046875,
-                                            372.81036376953125]
-
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             np.load('../../gemgis/data/Test1/array_rbf.npy')
-                         ])
-def test_extract_z_points_array(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=False, extent=[0, 972, 0, 1069])
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, np.ndarray)
-    assert all(gdf_new.geom_type == 'Point')
-
-    assert dem.ndim == 2
-    assert dem.shape == (1069, 972)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'Point')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
-                                            157.81229899479604, 191.31802803451436]
-    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
-                                            719.0939805375339]
-    assert gdf_new['Z'].head().tolist() == [387.2258761923539, 387.154888907343, 387.3960957643691, 387.5444087461885,
-                                            388.6688927116212]
-
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             np.load('../../gemgis/data/Test1/array_rbf.npy')
-                         ])
-def test_extract_z_points_array(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=True, extent=[0, 972, 0, 1069])
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, np.ndarray)
-    assert all(gdf_new.geom_type == 'Point')
-
-    assert dem.ndim == 2
-    assert dem.shape == (1069, 972)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'Point')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
-                                            157.81229899479604, 191.31802803451436]
-    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
-                                            719.0939805375339]
-    assert gdf_new['Z'].head().tolist() == [387.2258761923539, 387.154888907343, 387.3960957643691, 387.5444087461885,
-                                            388.6688927116212]
-
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             np.load('../../gemgis/data/Test1/array_rbf.npy')
-                         ])
-def test_extract_z_points_array(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=False, extent=[0, 972, 0, 1069])
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, np.ndarray)
-    assert all(gdf_new.geom_type == 'LineString')
-
-    assert dem.ndim == 2
-    assert dem.shape == (1069, 972)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'LineString')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
-                                            19.150128045807676,
-                                            27.79511673965105]
-    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
-                                            310.571692592952]
-    assert gdf_new['Z'].head().tolist() == [466.7501589231589, 468.49775671714633, 468.9434645548434,
-                                            469.09802654928296,
-                                            469.77232323980155]
-
-
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
-                         ])
-@pytest.mark.parametrize("dem",
-                         [
-                             np.load('../../gemgis/data/Test1/array_rbf.npy')
-                         ])
-def test_extract_z_values_points_array(gdf, dem):
-    from gemgis.vector import extract_z
-    gdf_new = extract_z(gdf, dem, inplace=True, extent=[0, 972, 0, 1069])
-
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert isinstance(dem, np.ndarray)
-    assert all(gdf_new.geom_type == 'LineString')
-
-    assert dem.ndim == 2
-    assert dem.shape == (1069, 972)
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'LineString')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
-                                            19.150128045807676,
-                                            27.79511673965105]
-    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
-                                            310.571692592952]
-    assert gdf_new['Z'].head().tolist() == [466.7501589231589, 468.49775671714633, 468.9434645548434,
-                                            469.09802654928296,
-                                            469.77232323980155]
 
 
 # Testing set_resolution
@@ -985,8 +582,6 @@ def test_sample_randomly_error(array):
         sample_randomly(array, [1000, 2069, 1000, 1972], seed=1.0)
     with pytest.raises(TypeError):
         sample_randomly(array, [1000, 2069, 1000, '1972'], seed=1)
-
-
 
 
 # Testing to_section_dict
@@ -1117,9 +712,9 @@ def test_convert_to_gempy_df_lines(gdf, dem):
                              rasterio.open('../../gemgis/data/Test1/raster1.tif')
                          ])
 def test_convert_to_gempy_df_lines_xyz(gdf, dem):
-    from gemgis.vector import extract_coordinates
+    from gemgis.vector import extract_xyz
     from gemgis.utils import convert_to_gempy_df
-    gdf_xyz = extract_coordinates(gdf, dem, inplace=True)
+    gdf_xyz = extract_xyz(gdf, dem)
     df = convert_to_gempy_df(gdf_xyz)
 
     assert dem.read(1).ndim == 2
@@ -1149,9 +744,9 @@ def test_convert_to_gempy_df_lines_xyz(gdf, dem):
                              rasterio.open('../../gemgis/data/Test1/raster1.tif')
                          ])
 def test_convert_to_gempy_df_points_xyz(gdf, dem):
-    from gemgis.vector import extract_coordinates
+    from gemgis.vector import extract_xyz
     from gemgis.utils import convert_to_gempy_df
-    gdf_xyz = extract_coordinates(gdf, dem, inplace=True)
+    gdf_xyz = extract_xyz(gdf, dem)
     df = convert_to_gempy_df(gdf_xyz)
 
     assert dem.read(1).ndim == 2
@@ -1183,7 +778,7 @@ def test_interpolate_raster_nearest(gdf):
     from gemgis.vector import interpolate_raster
     from gemgis.vector import extract_xy
 
-    gdf_xyz = extract_xy(gdf, inplace=False)
+    gdf_xyz = extract_xy(gdf)
     raster = interpolate_raster(gdf_xyz, method='nearest')
 
     assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
@@ -1200,7 +795,7 @@ def test_interpolate_raster_linear(gdf):
     from gemgis.vector import interpolate_raster
     from gemgis.vector import extract_xy
 
-    gdf_xyz = extract_xy(gdf, inplace=False)
+    gdf_xyz = extract_xy(gdf)
     raster = interpolate_raster(gdf_xyz, method='linear')
 
     assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
@@ -1217,7 +812,7 @@ def test_interpolate_raster_cubic(gdf):
     from gemgis.vector import interpolate_raster
     from gemgis.vector import extract_xy
 
-    gdf_xyz = extract_xy(gdf, inplace=False)
+    gdf_xyz = extract_xy(gdf)
     raster = interpolate_raster(gdf_xyz, method='cubic')
 
     assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
@@ -1234,7 +829,7 @@ def test_interpolate_raster_rbf(gdf):
     from gemgis.vector import interpolate_raster
     from gemgis.vector import extract_xy
 
-    gdf_xyz = extract_xy(gdf, inplace=False)
+    gdf_xyz = extract_xy(gdf)
     raster = interpolate_raster(gdf_xyz, method='rbf')
 
     assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
@@ -1251,7 +846,7 @@ def test_interpolate_raster_error(gdf):
     from gemgis.vector import interpolate_raster
     from gemgis.vector import extract_xy
 
-    gdf_xyz = extract_xy(gdf, inplace=False)
+    gdf_xyz = extract_xy(gdf)
     with pytest.raises(TypeError):
         interpolate_raster([gdf_xyz], method='linear')
     with pytest.raises(TypeError):
@@ -1266,7 +861,7 @@ def test_interpolate_raster_rbf_samples(gdf):
     from gemgis.vector import interpolate_raster
     from gemgis.vector import extract_xy
 
-    gdf_xyz = extract_xy(gdf, inplace=False)
+    gdf_xyz = extract_xy(gdf)
     raster = interpolate_raster(gdf_xyz, method='rbf', n=30)
 
     assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
@@ -1824,9 +1419,9 @@ def test_save_raster_as_tiff(raster):
                          ])
 def test_plot_points_3d(gdf, dem):
     from gemgis.visualization import plot_points_3d
-    from gemgis.vector import extract_z
+    from gemgis.vector import extract_xyz
 
-    gdf = extract_z(gdf, dem)
+    gdf = extract_xyz(gdf, dem)
     p = pv.Plotter(notebook=True)
 
     assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
@@ -1854,9 +1449,9 @@ def test_plot_points_3d(gdf, dem):
                          ])
 def test_plot_points_3d_error(gdf, dem):
     from gemgis.visualization import plot_points_3d
-    from gemgis.vector import extract_z
+    from gemgis.vector import extract_xyz
 
-    gdf = extract_z(gdf, dem)
+    gdf = extract_xyz(gdf, dem)
     p = pv.Plotter(notebook=True)
 
     assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
@@ -2053,71 +1648,6 @@ def test_plot_contours_3d_error(lines):
         plot_contours_3d(lines, [p], color='red')
     with pytest.raises(TypeError):
         plot_contours_3d(lines, p, color=['red'])
-
-
-# Testing clip_vector_data_by_extent
-###########################################################
-@pytest.mark.parametrize("points",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/randompoints1.shp')
-                         ])
-def test_clip_vector_data_by_extent(points):
-    from gemgis.vector import clip_by_extent
-
-    gdf = clip_by_extent(points, [0, 1069, 0, 972])
-
-    assert len(points) == 50
-    assert len(gdf) == 24
-
-    assert points.bounds.min()[0] == -471.08143156369886
-    assert points.bounds.min()[1] == -203.33821461472303
-    assert points.bounds.min()[2] == -471.08143156369886
-    assert points.bounds.min()[3] == -203.33821461472303
-
-    assert points.bounds.max()[0] == 1321.8130710900514
-    assert points.bounds.max()[1] == 1347.6776138284376
-    assert points.bounds.max()[2] == 1321.8130710900514
-    assert points.bounds.max()[3] == 1347.6776138284376
-
-    assert gdf.bounds.min()[0] == 100.59482325004632
-    assert gdf.bounds.min()[1] == 73.45078370873489
-    assert gdf.bounds.min()[2] == 100.59482325004632
-    assert gdf.bounds.min()[3] == 73.45078370873489
-
-    assert gdf.bounds.max()[0] == 925.866703136033
-    assert gdf.bounds.max()[1] == 946.2088865304491
-    assert gdf.bounds.max()[2] == 925.866703136033
-    assert gdf.bounds.max()[3] == 946.2088865304491
-
-    assert isinstance(points, gpd.geodataframe.GeoDataFrame)
-    assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
-    assert not {'X', 'Y', 'Z'}.issubset(points.columns)
-    assert {'X', 'Y'}.issubset(gdf.columns)
-
-    assert gdf is not points
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-    assert points.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(points.geom_type == 'Point')
-
-    assert 'geometry' in gdf
-    assert all(gdf.geom_type == 'Point')
-
-
-@pytest.mark.parametrize("points",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/randompoints1.shp')
-                         ])
-def test_clip_vector_data_by_extent_error(points):
-    from gemgis.vector import clip_by_extent
-
-    with pytest.raises(TypeError):
-        clip_by_extent([points], [0, 1069, 0, 972])
-    with pytest.raises(TypeError):
-        clip_by_extent(points, (0, 1069, 0, 972))
 
 
 # Testing clip_vector_data_by_shape
@@ -2389,7 +1919,7 @@ def test_sample_orientations_from_raster_point(dem):
 
     extent = set_extent(0, 972, 0, 1069)
 
-    orientations = sample_orientations(dem, extent, points=[500, 500], formation='surface')
+    orientations = sample_orientations(dem.read(1), extent, points=[500, 500], formation='surface')
 
     assert isinstance(orientations, pd.DataFrame)
     assert {'X', 'Y', 'Z', 'formation', 'dip', 'azimuth', 'polarity'}.issubset(orientations.columns)
@@ -2406,7 +1936,7 @@ def test_sample_orientations_from_raster_points(dem):
 
     extent = set_extent(0, 972, 0, 1069)
 
-    orientations = sample_orientations(dem, extent, points=[[500, 500], [600, 600]], formation='surface')
+    orientations = sample_orientations(dem.read(1), extent, points=[[500, 500], [600, 600]], formation='surface')
 
     assert isinstance(orientations, pd.DataFrame)
     assert {'X', 'Y', 'Z', 'formation', 'dip', 'azimuth', 'polarity'}.issubset(orientations.columns)
@@ -2423,7 +1953,7 @@ def test_sample_orientations_from_raster_points3(dem):
 
     extent = set_extent(0, 972, 0, 1069)
 
-    orientations = sample_orientations(dem, extent, points=[[500, 500], [600, 600], [700, 700]],
+    orientations = sample_orientations(dem.read(1), extent, points=[[500, 500], [600, 600], [700, 700]],
                                        formation='surface')
 
     assert isinstance(orientations, pd.DataFrame)
@@ -2482,7 +2012,7 @@ def test_sample_interfaces_from_raster_point(dem):
 
     extent = set_extent(0, 972, 0, 1069)
 
-    interfaces = sample_interfaces(dem, extent, points=[500, 500], formation='surface')
+    interfaces = sample_interfaces(dem.read(1), extent, points=[500, 500], formation='surface')
 
     assert isinstance(interfaces, pd.DataFrame)
     assert {'X', 'Y', 'Z', 'formation'}.issubset(interfaces.columns)
@@ -2499,7 +2029,7 @@ def test_sample_interfaces_from_raster_points(dem):
 
     extent = set_extent(0, 972, 0, 1069)
 
-    interfaces = sample_interfaces(dem, extent, points=[[500, 500], [600, 600]], formation='surface')
+    interfaces = sample_interfaces(dem.read(1), extent, points=[[500, 500], [600, 600]], formation='surface')
 
     assert isinstance(interfaces, pd.DataFrame)
     assert {'X', 'Y', 'Z', 'formation'}.issubset(interfaces.columns)
@@ -2516,12 +2046,12 @@ def test_sample_interfaces_from_raster_points3(dem):
 
     extent = set_extent(0, 972, 0, 1069)
 
-    interfaces = sample_interfaces(dem, extent, points=[[500, 500], [600, 600], [700, 700]],
+    interfaces = sample_interfaces(dem.read(1), extent, points_x=[500, 600, 700], points_y=[500, 600, 700],
                                    formation='surface')
 
     assert isinstance(interfaces, pd.DataFrame)
     assert {'X', 'Y', 'Z', 'formation'}.issubset(interfaces.columns)
-    assert len(interfaces) == 3
+    assert len(interfaces) == 10
 
 
 @pytest.mark.parametrize("dem",
@@ -2535,13 +2065,13 @@ def test_sample_interfaces_from_raster_error(dem):
     extent = set_extent(0, 972, 0, 1069)
 
     with pytest.raises(TypeError):
-        sample_interfaces([dem], extent, points=[[500, 500], [600, 600], [700, 700]], formation='surface')
-    with pytest.raises(ValueError):
-        sample_interfaces(dem, [extent], points=[[500, 500], [600, 600], [700, 700]], formation='surface')
+        sample_interfaces([dem], extent, points_x=[[500, 500], [600, 600], [700, 700]], points_y=[[500, 500], [600, 600], [700, 700]], formation='surface')
     with pytest.raises(TypeError):
-        sample_interfaces(dem, extent, points=([500, 500], [600, 600], [700, 700]), formation='surface')
+        sample_interfaces(dem, [extent], points_x=[[500, 500], [600, 600], [700, 700]], points_y=[[500, 500], [600, 600], [700, 700]], formation='surface')
     with pytest.raises(TypeError):
-        sample_interfaces(dem, extent, points=[[500, 500], [600, 600], [700, 700]], formation=['surface'])
+        sample_interfaces(dem, 'extent', points_x=([500, 500], [600, 600], [700, 700]), points_y=[[500, 500], [600, 600], [700, 700]], formation='surface')
+    with pytest.raises(TypeError):
+        sample_interfaces(dem, extent, points_x=[[500, 500], [600, 600], [700, 700]], points_y=[[500, 500], [600, 600], [700, 700]], formation=['surface'])
     with pytest.raises(TypeError):
         sample_interfaces([dem], extent, formation='surface')
 
@@ -2670,7 +2200,7 @@ def test_calculate_orientations(points):
 # Testing create_surface_color_dict
 ###########################################################
 
-def test_create_surface_color_dict():
+def test_create_surface_color_dict1():
     from gemgis.utils import create_surface_color_dict
 
     surface_color_dict = create_surface_color_dict('../../gemgis/data/Test1/style1.qml')
@@ -2679,7 +2209,7 @@ def test_create_surface_color_dict():
     assert surface_color_dict == {'Sand1': '#b35a2a', 'Sand2': '#b35a2a', 'Ton': '#525252'}
 
 
-def test_create_surface_color_dict_error():
+def test_create_surface_color_dict_error2():
     from gemgis.utils import create_surface_color_dict
 
     with pytest.raises(TypeError):
@@ -2895,7 +2425,7 @@ def test_interpolate_strike_lines(gdf):
 
     assert isinstance(lines, gpd.geodataframe.GeoDataFrame)
     assert lines.crs == 'EPSG:4326'
-    assert len(lines) == 16
+    assert len(lines) == 33
     assert {'X', 'Y', 'Z'}.issubset(lines.columns)
 
 
@@ -2938,10 +2468,10 @@ def test_load_wfs():
                          ])
 def test_show_number_of_data_points(interfaces, orientations, dem):
     from gemgis.utils import show_number_of_data_points
-    from gemgis.vector import extract_coordinates
+    from gemgis.vector import extract_xyz
 
-    interfaces_coords = extract_coordinates(interfaces, dem, extent=[-0.0, 972.0, -0.0, 1069.0])
-    orientations_coords = extract_coordinates(orientations, dem, extent=[-0.0, 972.0, -0.0, 1069.0])
+    interfaces_coords = extract_xyz(interfaces, dem, extent=[-0.0, 972.0, -0.0, 1069.0])
+    orientations_coords = extract_xyz(orientations, dem, extent=[-0.0, 972.0, -0.0, 1069.0])
 
     geo_model = gp.create_model('Test')
 
@@ -3148,110 +2678,6 @@ def test_polygons_to_linestrings(gdf):
     assert isinstance(gdf_linestrings, gpd.geodataframe.GeoDataFrame)
     assert gdf_linestrings.crs == 'EPSG:4647'
     assert len(gdf_linestrings) == 848
-
-
-# Testing extract_xy dropping columns
-###########################################################
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
-                         ])
-def test_extract_xy_drop_id(gdf):
-    from gemgis.vector import extract_xy
-
-    gdf_new = extract_xy(gdf, inplace=False)
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert all(gdf.geom_type == 'Point')
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf_new, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'Point')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y'}.issubset(gdf_new.columns)
-
-    # Assert if values are correct
-    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
-                                            157.81229899479604, 191.31802803451436]
-    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049,
-                                            615.9994296460927,
-                                            719.0939805375339]
-
-    assert not {'id'}.issubset(gdf_new.columns)
-
-
-# Testing extract_xy dropping columns 2
-###########################################################
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
-                         ])
-def test_extract_xy_drop_index(gdf):
-    from gemgis.vector import extract_xy
-
-    gdf_new = extract_xy(gdf, inplace=False)
-    # Assert type on input
-    assert isinstance(gdf, gpd.GeoDataFrame)
-    assert 'geometry' in gdf
-    assert all(gdf.geom_type == 'LineString')
-
-    # Assert CRS
-    assert gdf.crs == 'EPSG:4326'
-
-    # Assert if columns are already in input gdf
-    assert not {'X', 'Y'}.issubset(gdf.columns)
-
-    # Assert type of output
-    assert isinstance(gdf_new, gpd.GeoDataFrame)
-    assert gdf is not gdf_new
-
-    # Assert CRS
-    assert gdf_new.crs == 'EPSG:4326'
-
-    # Assert Type of shape file
-    assert all(gdf_new.geom_type == 'LineString')
-
-    # Assert if columns are in gdf_new
-    assert {'X', 'Y'}.issubset(gdf_new.columns)
-
-    assert not {'index'}.issubset(gdf_new.columns)
-
-
-# Testing extract_xy for MultiLineStrings and LineStrings
-###########################################################
-@pytest.mark.parametrize("gdf",
-                         [
-                             gpd.read_file('../../gemgis/data/tutorials/tutorial13/GeologicalMapAachen.shp')
-                         ])
-def test_extract_xy(gdf):
-    from gemgis.vector import polygons_to_linestrings
-    from gemgis.vector import extract_xy
-
-    gdf_linestrings = polygons_to_linestrings(gdf)
-
-    gdf_linestrings_xy = extract_xy(gdf_linestrings)
-
-    assert isinstance(gdf_linestrings_xy, gpd.geodataframe.GeoDataFrame)
-
-    assert gdf_linestrings_xy.crs == 'EPSG:4647'
-
-    assert not {'id'}.issubset(gdf_linestrings_xy.columns)
-    assert not {'index'}.issubset(gdf_linestrings_xy.columns)
-    assert {'X', 'Y', 'points', 'geometry'}.issubset(gdf_linestrings_xy.columns)
 
 
 # Testing calculate_strike
@@ -3493,10 +2919,10 @@ def test_extract_borehole(interfaces, orientations, dem):
     geo_data.set_extent(-0.0, 972.0, -0.0, 1069.0, 300, 800)
     geo_data.set_resolution(50, 50, 50)
 
-    interfaces_coords = gg.vector.extract_coordinates(interfaces, dem, extent=geo_data.extent)
+    interfaces_coords = gg.vector.extract_xyz(interfaces, dem, extent=geo_data.extent)
     geo_data.to_gempy_df(interfaces_coords, 'interfaces')
 
-    orientations_coords = gg.vector.extract_coordinates(orientations, dem, extent=geo_data.extent)
+    orientations_coords = gg.vector.extract_xyz(orientations, dem, extent=geo_data.extent)
     geo_data.to_gempy_df(orientations_coords, 'orientations')
 
     geo_data.stack = {"Strat_Series": ('Sand1', 'Ton')}
