+++ tests/test_vector.py
@@ -2155,3 +2155,576 @@ def test_extract_coordinates_countours(gdf):
 
     assert 'geometry' in gdf
     assert all(gdf.geom_type == 'LineString')
+
+
+# Testing extract_z
+###########################################################
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
+                         ])
+def test_extract_z_points(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem)
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, rasterio.io.DatasetReader)
+    assert all(gdf_new.geom_type == 'Point')
+
+    assert dem.read(1).ndim == 2
+    assert dem.read(1).shape == (275, 250)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+    assert dem.crs == {'init': 'epsg:4326'}
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
+                                            157.81229899479604, 191.31802803451436]
+    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
+                                            719.0939805375339]
+    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
+                                            719.0939805375339]
+
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
+                         ])
+def test_extract_z_points_inplace(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem)
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, rasterio.io.DatasetReader)
+    assert all(gdf_new.geom_type == 'Point')
+
+    assert dem.read(1).ndim == 2
+    assert dem.read(1).shape == (275, 250)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+    assert dem.crs == {'init': 'epsg:4326'}
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
+                                            157.81229899479604, 191.31802803451436]
+    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
+                                            719.0939805375339]
+    assert gdf_new['Z'].head().tolist() == [364.994873046875, 400.3435974121094, 459.54931640625, 525.6910400390625,
+                                            597.6325073242188]
+
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
+                         ])
+def test_extract_z_lines_inplace(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem, inplace=False)
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, rasterio.io.DatasetReader)
+    assert all(gdf_new.geom_type == 'LineString')
+
+    assert dem.read(1).ndim == 2
+    assert dem.read(1).shape == (275, 250)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+    assert dem.crs == {'init': 'epsg:4326'}
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'LineString')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
+                                            19.150128045807676, 27.79511673965105]
+    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
+                                            310.571692592952]
+    assert gdf_new['Z'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
+                                            310.571692592952]
+
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             rasterio.open('../../gemgis/data/Test1/raster1.tif')
+                         ])
+def test_extract_z_lines_inplace(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem)
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, rasterio.io.DatasetReader)
+    assert all(gdf_new.geom_type == 'Point')
+
+    assert dem.read(1).ndim == 2
+    assert dem.read(1).shape == (275, 250)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+    assert dem.crs == {'init': 'epsg:4326'}
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
+                                            19.150128045807676, 27.79511673965105]
+    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
+                                            310.571692592952]
+    assert gdf_new['Z'].head().tolist() == [353.9727783203125, 359.03631591796875, 364.28497314453125, 364.994873046875,
+                                            372.81036376953125]
+
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             np.load('../../gemgis/data/Test1/array_rbf.npy')
+                         ])
+def test_extract_z_points_array(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem, inplace=False, extent=[0, 972, 0, 1069])
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, np.ndarray)
+    assert all(gdf_new.geom_type == 'Point')
+
+    assert dem.ndim == 2
+    assert dem.shape == (1069, 972)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
+                                            157.81229899479604, 191.31802803451436]
+    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
+                                            719.0939805375339]
+    assert gdf_new['Z'].head().tolist() == [387.2258761923539, 387.154888907343, 387.3960957643691, 387.5444087461885,
+                                            388.6688927116212]
+
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             np.load('../../gemgis/data/Test1/array_rbf.npy')
+                         ])
+def test_extract_z_points_array(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem, inplace=True, extent=[0, 972, 0, 1069])
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, np.ndarray)
+    assert all(gdf_new.geom_type == 'Point')
+
+    assert dem.ndim == 2
+    assert dem.shape == (1069, 972)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
+                                            157.81229899479604, 191.31802803451436]
+    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049, 615.9994296460927,
+                                            719.0939805375339]
+    assert gdf_new['Z'].head().tolist() == [387.2258761923539, 387.154888907343, 387.3960957643691, 387.5444087461885,
+                                            388.6688927116212]
+
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             np.load('../../gemgis/data/Test1/array_rbf.npy')
+                         ])
+def test_extract_z_points_array(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem, extent=[0, 972, 0, 1069])
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, np.ndarray)
+    assert all(gdf_new.geom_type == 'Point')
+
+    assert dem.ndim == 2
+    assert dem.shape == (1069, 972)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
+                                            19.150128045807676,
+                                            27.79511673965105]
+    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
+                                            310.571692592952]
+    assert gdf_new['Z'].head().tolist() == [466.7501589231589, 468.49775671714633, 468.9434645548434,
+                                            469.09802654928296,
+                                            469.77232323980155]
+
+
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
+                         ])
+@pytest.mark.parametrize("dem",
+                         [
+                             np.load('../../gemgis/data/Test1/array_rbf.npy')
+                         ])
+def test_extract_z_values_points_array(gdf, dem):
+    from gemgis.vector import extract_xyz
+    gdf_new = extract_xyz(gdf, dem, extent=[0, 972, 0, 1069])
+
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert isinstance(dem, np.ndarray)
+    assert all(gdf_new.geom_type == 'Point')
+
+    assert dem.ndim == 2
+    assert dem.shape == (1069, 972)
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y', 'Z'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y', 'Z'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [0.256327195431048, 10.59346813871597, 17.134940141888464,
+                                            19.150128045807676,
+                                            27.79511673965105]
+    assert gdf_new['Y'].head().tolist() == [264.86214748436396, 276.73370778641777, 289.089821570188, 293.313485355882,
+                                            310.571692592952]
+    assert gdf_new['Z'].head().tolist() == [466.7501589231589, 468.49775671714633, 468.9434645548434,
+                                            469.09802654928296,
+                                            469.77232323980155]
+
+
+# Testing clip_vector_data_by_extent
+###########################################################
+@pytest.mark.parametrize("points",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/randompoints1.shp')
+                         ])
+def test_clip_vector_data_by_extent(points):
+    from gemgis.vector import clip_by_bbox
+
+    gdf = clip_by_bbox(points, [0, 1069, 0, 972])
+
+    assert len(points) == 50
+    assert len(gdf) == 24
+
+    assert points.bounds.min()[0] == -471.08143156369886
+    assert points.bounds.min()[1] == -203.33821461472303
+    assert points.bounds.min()[2] == -471.08143156369886
+    assert points.bounds.min()[3] == -203.33821461472303
+
+    assert points.bounds.max()[0] == 1321.8130710900514
+    assert points.bounds.max()[1] == 1347.6776138284376
+    assert points.bounds.max()[2] == 1321.8130710900514
+    assert points.bounds.max()[3] == 1347.6776138284376
+
+    assert gdf.bounds.min()[0] == 100.59482325004632
+    assert gdf.bounds.min()[1] == 73.45078370873489
+    assert gdf.bounds.min()[2] == 100.59482325004632
+    assert gdf.bounds.min()[3] == 73.45078370873489
+
+    assert gdf.bounds.max()[0] == 925.866703136033
+    assert gdf.bounds.max()[1] == 946.2088865304491
+    assert gdf.bounds.max()[2] == 925.866703136033
+    assert gdf.bounds.max()[3] == 946.2088865304491
+
+    assert isinstance(points, gpd.geodataframe.GeoDataFrame)
+    assert isinstance(gdf, gpd.geodataframe.GeoDataFrame)
+    assert not {'X', 'Y', 'Z'}.issubset(points.columns)
+    assert {'X', 'Y'}.issubset(gdf.columns)
+
+    assert gdf is not points
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+    assert points.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(points.geom_type == 'Point')
+
+    assert 'geometry' in gdf
+    assert all(gdf.geom_type == 'Point')
+
+
+@pytest.mark.parametrize("points",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/randompoints1.shp')
+                         ])
+def test_clip_vector_data_by_extent_error(points):
+    from gemgis.vector import clip_by_bbox
+
+    with pytest.raises(TypeError):
+        clip_by_bbox([points], [0, 1069, 0, 972])
+    with pytest.raises(TypeError):
+        clip_by_bbox(points, (0, 1069, 0, 972))
+
+
+# Testing extract_xy dropping columns
+###########################################################
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1.shp')
+                         ])
+def test_extract_xy_drop_id(gdf):
+    from gemgis.vector import extract_xy
+
+    gdf_new = extract_xy(gdf)
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert all(gdf.geom_type == 'Point')
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf_new, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y'}.issubset(gdf_new.columns)
+
+    # Assert if values are correct
+    assert gdf_new['X'].head().tolist() == [19.150128045807676, 61.93436666575576, 109.35786007581868,
+                                            157.81229899479604, 191.31802803451436]
+    assert gdf_new['Y'].head().tolist() == [293.313485355882, 381.4593263680641, 480.9455679783049,
+                                            615.9994296460927,
+                                            719.0939805375339]
+
+    assert not {'id'}.issubset(gdf_new.columns)
+
+
+# Testing extract_xy dropping columns 2
+###########################################################
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/Test1/interfaces1_lines.shp')
+                         ])
+def test_extract_xy_drop_index(gdf):
+    from gemgis.vector import extract_xy
+
+    gdf_new = extract_xy(gdf)
+    # Assert type on input
+    assert isinstance(gdf, gpd.GeoDataFrame)
+    assert 'geometry' in gdf
+    assert all(gdf.geom_type == 'LineString')
+
+    # Assert CRS
+    assert gdf.crs == 'EPSG:4326'
+
+    # Assert if columns are already in input gdf
+    assert not {'X', 'Y'}.issubset(gdf.columns)
+
+    # Assert type of output
+    assert isinstance(gdf_new, gpd.GeoDataFrame)
+    assert gdf is not gdf_new
+
+    # Assert CRS
+    assert gdf_new.crs == 'EPSG:4326'
+
+    # Assert Type of shape file
+    assert all(gdf_new.geom_type == 'Point')
+
+    # Assert if columns are in gdf_new
+    assert {'X', 'Y'}.issubset(gdf_new.columns)
+
+    assert not {'index'}.issubset(gdf_new.columns)
+
+
+# Testing extract_xy for MultiLineStrings and LineStrings
+###########################################################
+@pytest.mark.parametrize("gdf",
+                         [
+                             gpd.read_file('../../gemgis/data/tutorials/tutorial13/GeologicalMapAachen.shp')
+                         ])
+def test_extract_xy_multilinestrings2(gdf):
+    from gemgis.vector import polygons_to_linestrings
+    from gemgis.vector import extract_xy
+
+    gdf_linestrings = polygons_to_linestrings(gdf)
+
+    gdf_linestrings_xy = extract_xy(gdf_linestrings)
+
+    assert isinstance(gdf_linestrings_xy, gpd.geodataframe.GeoDataFrame)
+
+    assert gdf_linestrings_xy.crs == 'EPSG:4647'
+
+    assert not {'id'}.issubset(gdf_linestrings_xy.columns)
+    assert not {'index'}.issubset(gdf_linestrings_xy.columns)
+    assert {'X', 'Y', 'geometry'}.issubset(gdf_linestrings_xy.columns)
+
